requirements.txt
# AI-Powered Speech Guidance Self-Assistant
# Raspberry Pi 4 + ReSpeaker 2-Mics Pi HAT v2.0 + Pi Camera Rev 1.3

# --- Core ---
PyYAML>=6.0
numpy>=1.24.0

# --- Web / Auth ---
flask-login>=0.6.0

# --- Audio (ReSpeaker 2-Mics Pi HAT) ---
pyaudio>=0.2.13
webrtcvad>=2.0.10
sounddevice>=0.4.6

# --- Speech-to-Text ---
openai-whisper>=20231117
SpeechRecognition>=3.10.0

# --- AI / LLM ---
ollama>=0.3.0
openai>=1.12.0

# --- Text-to-Speech ---
pyttsx3>=2.90
gTTS>=2.5.0

# --- Camera ---
picamera2>=0.3.12

# --- LED Control (ReSpeaker APA102) ---
spidev>=3.6
gpiod>=2.0

# --- Web UI (optional) ---
flask>=3.0.0
flask-socketio>=5.3.0

# --- Utilities ---
Pillow>=10.0.0
requests>=2.31.0
